+++ program/game/TEPPO.cpp
@@ -20,7 +20,7 @@ TEPPO::TEPPO(tpr::Vector2 pos_, int angle) {
 	r = 14;
 
 	sound = GMp->loadSoundMem("sound/shot.mp3");
-	ChangeVolumeSoundMem(255 * 0.4, sound);
+	ChangeVolumeSoundMem((int)(255 * 0.4), sound);
 
 	//=============================================================================
 	//デバッグ用
